requirements.txt
discord.py==1.5.1
sqlalchemy==1.3.22
psycopg2-binary==2.8.6
numpy==1.19.5